Restrict operator syntax
The relaxed operator parsing broke in session types like

  ?+-+-A.end

However, it is unclear if this should even parse correctly or if we want
to require parentheses around each +/- operator application. Keep in
mind that this is not a normal use case but a rather quite strange of
way of writing

  ?A.end

For now we only parse the builtin operators.