Slight lexer improvement
By using an accumulator we reduce the amount of stack space required in
the lexer. Additionally, we eliminate the `trim` post-processing step
which required two reversals of the whole list of tokens.

As a small improvement in the parser, multiple newline tokens are now
compressed in the lexer which eliminates the need of one parsing rule.