Build container with UTF-8 encoding baked in
This comes from a complicated interaction between terminal encoding,
STDOUT encoding and the docker boundary.

Haskell takes the encoding of its STDOUT from the systems default
encoding. We would like to use UTF-8, but this is not guaranteed. So we
use a fall-back which replaces (in the worst case) any non ASCII
characters with a question mark.

When run from our docker container without any explicit additional
information the default encoding is "POSIX" (=ASCII). But this encoding
has nothing to do with the user's terminal's encoding. So we will now
default to using a UTF-8 encoding in the docker images. But this can be
overriden by the user when running using `--env LANG=...`.